Update draft-azimov-sidrops-aspa-verification.xml
1. I have added two new sections:
        5.4.  Upstream AS Path Validation Algorithm
        5.5.  Downstream AS Path Validation 
They address Alexander's requests for refinements based on "Sriram's findings". These also take care of AS_SET scenarios. The existing algorithms (Sec. 5.1, 5.2, 5.3) can be replaced if these are accepted.
They also address transparent/non-transparent RS inclusion without the need to have a separate algorithms for RS. The downstream algorithm is good for RS also. In the case of transparent RS, the RS ASN is added to the AS path (for validation only).